bishop/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "bishop frontend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.12.4",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
